Split tests into single and weighted
