+++ jest.config.e2e.js
@@ -1,3 +1,4 @@
 module.exports = {
-  testTimeout: 240000
+  testTimeout: 240000,
+  testRegex: '((\\.|/)(test|spec))\\.[jt]sx?$'
 }
