+++ package.json
@@ -18,7 +18,7 @@
     "standard": "ts-standard --fix",
     "test": "jest",
     "test:ci": "jest --ci --coverage --forceExit --maxWorkers=4",
-    "test:e2e": "jest --config ./jest.config.e2e.js .github/workflows/e2e/test.js",
+    "test:e2e": "jest --runInBand --config ./jest.config.e2e.js .github/workflows/e2e/",
     "all": "npm run build && npm run standard && npm run test",
     "package": "npm run pack -w ./packages/salmon-lambda-functions"
   },
